Fix unread chat messages badge on the CallWithChatComposite being cleared when the call goes on hold (#3332)
