+++ change/@azure-communication-react-ff67bea7-5ec4-4ff7-851c-b829cebc71cc.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Fix unread chat messages badge on the CallWithChatComposite being cleared when the call goes on hold",
+  "packageName": "@azure/communication-react",
+  "email": "2684369+JamesBurnside@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
